feature/story-1: TestCalculateTaxPayableAndRefund
